+++ collatz.c
@@ -5,21 +5,21 @@
 
 #include "cache.h"
 
-int32_t even(int32_t n) {
+uint64_t even(uint64_t n) {
 	// printf("even %d\n", n);
 	// n / 2
 	return n / 2;
 }
 
-int32_t odd(int32_t n) {
+uint64_t odd(uint64_t n) {
 	// printf("odd %d\n", n);
 	// 3n + 1
 	return (n * 3) + 1;
 }
 
-int32_t step(int32_t n) {
-	if (n > ((INT_MAX-1) / 3)) {
-		fprintf(stderr, "integer overflow at %d\n", n);
+uint64_t step(uint64_t n) {
+	if (n > ((INT64_MAX-1) / 3)) {
+		fprintf(stderr, "integer overflow at %lld\n", n);
 		exit(2);
 	}
 	if ((n & 1) == 1) {
@@ -28,7 +28,7 @@ int32_t step(int32_t n) {
 	return even(n);
 }
 
-int32_t collatz(int32_t n, struct CacheEntry cache[]) {
+uint64_t collatz(uint64_t n, struct CacheEntry cache[]) {
 	int64_t count;
 	for (count=0; n>1; ++count) {
 		for (int i=0; i<MAX_CACHE; i++) {
