+++ main.c
@@ -8,18 +8,18 @@
 #include "collatz.h"
 #include "cache.h"
 
-struct CacheEntry main1(int start, int end);
+struct CacheEntry main1(uint64_t start, uint64_t end);
 void* thread_main(void *arg);
 
 struct WorkEntry {
-	int upper;
-	int lower;
+	uint64_t upper;
+	uint64_t lower;
 	struct CacheEntry result;
 };
 
 int main(int argc, char **argv) {
-	int n;
-	int work_set_size;
+	uint64_t n;
+	uint64_t work_set_size;
 
 	if (argc != 2) {
 		printf("Usage: %s [int]\n", argv[0]);
@@ -36,11 +36,11 @@ int main(int argc, char **argv) {
 	for (int i=0; i<MAX_THREADS; i++) {
 		results[i].lower = (work_set_size * i)+1;
 		results[i].upper   = (work_set_size * (i+1));
-		// printf("%d vs %d\n", results[i].upper, n);
+		// printf("%lld vs %lld\n", results[i].upper, n);
 		if (results[i].upper > n) {
 			results[i].upper = n;
 		}
-		// printf("work upper:%d lower:%d\n", results[i].upper, results[i].lower);
+		// printf("work upper:%lld lower:%lld\n", results[i].upper, results[i].lower);
 
 		// results[i] = main1(upper, lower);
 		if(pthread_create(&workers[i], NULL, thread_main, &results[i])) {
@@ -58,7 +58,7 @@ int main(int argc, char **argv) {
 			result = results[i].result;
 		}
 	}
-	printf("result: number:%d length:%d\n", result.number, result.length);
+	printf("result: number:%lld length:%lld\n", result.number, result.length);
 }
 
 void* thread_main(void *arg) {
@@ -69,12 +69,12 @@ void* thread_main(void *arg) {
 	return NULL;
 }
 
-struct CacheEntry main1(int start, int end) {
+struct CacheEntry main1(uint64_t start, uint64_t end) {
 	int64_t tmp = 0;
 	int64_t longestChain = 0;
-	int result = 0;
+	uint64_t result = 0;
 	struct CacheEntry *cache;
-	int cacheIndex = 0;
+	uint64_t cacheIndex = 0;
 	struct CacheEntry newEntry;
 
 
@@ -84,7 +84,7 @@ struct CacheEntry main1(int start, int end) {
 		cache[i].length = 0;
 	}
 
-	for (int i=start; i>=end; --i) {
+	for (uint64_t i=start; i>=end; --i) {
 		tmp = collatz(i, cache);
 		if (tmp > longestChain) {
 			result = i;
@@ -97,7 +97,7 @@ struct CacheEntry main1(int start, int end) {
 			cacheIndex = 0;
 		}
 	}
-	// printf("result %d, count %lld\n", result, longestChain);
+	// printf("result %lld, count %lld\n", result, longestChain);
 
 	free(cache);
